refactor profile dropdown
